requirements.txt
streamlit
SpeechRecognition
openai
pillow
pytesseract
mss
numpy
python-dotenv
pyaudio
